Update application/language/nl/validation.php
The use of 'Het' is not always correct, sometimes it should be 'De', so the sentence would look weird. Probably better to remove it.

Also, the 'exists' string stated it already exists, but should state it doesn't exist! Very important difference ;)

Also changed the formatting errors of email/url to be more specific.